refactor(tests): move screen tracking tests to dedicated test file
- Reverted MixpanelBaseTests.swift to original state
- Created MixpanelScreenTrackingTests.swift with all screenView/screenLeave tests
- Tests now properly use removeDBfile for cleanup
