+++ MixpanelDemo/MixpanelDemoTests/MixpanelBaseTests.swift
@@ -49,9 +49,9 @@ class MixpanelBaseTests: XCTestCase, MixpanelDelegate {
   func dbFilePath(_ token: String? = nil) -> String {
     let manager = FileManager.default
     #if os(iOS)
-    let url = manager.urls(for: .libraryDirectory, in: .userDomainMask).last
+      let url = manager.urls(for: .libraryDirectory, in: .userDomainMask).last
     #else
-    let url = manager.urls(for: .cachesDirectory, in: .userDomainMask).last
+      let url = manager.urls(for: .cachesDirectory, in: .userDomainMask).last
     #endif  // os(iOS)
     guard let apiToken = token else {
       return ""
@@ -119,91 +119,6 @@ class MixpanelBaseTests: XCTestCase, MixpanelDelegate {
     waitForTrackingQueue(mixpanel)
     mixpanel.flush()
     waitForTrackingQueue(mixpanel)
-  }
-
-  func testScreenView() {
-    let testMixpanel = Mixpanel.initialize(token: randomId(), trackAutomaticEvents: false)
-    testMixpanel.screenView(screenName: "HomeScreen", properties: ["extra_prop": "extra_value"])
-
-    waitForTrackingQueue(testMixpanel)
-
-    let events = eventQueue(token: testMixpanel.apiToken)
-    XCTAssertEqual(events.count, 1)
-
-    let event = events.first
-    XCTAssertEqual(event?["event"] as? String, "$mp_page_view")
-
-    let props = event?["properties"] as? InternalProperties
-    XCTAssertEqual(props?["current_page_title"] as? String, "HomeScreen")
-    XCTAssertEqual(props?["extra_prop"] as? String, "extra_value")
-    XCTAssertNotNil(props?["$screen_height"])
-
-    removeDBfile(testMixpanel.apiToken)
-  }
-
-  func testScreenViewWithoutProperties() {
-    let testMixpanel = Mixpanel.initialize(token: randomId(), trackAutomaticEvents: false)
-    testMixpanel.screenView(screenName: "HomeScreen")
-
-    waitForTrackingQueue(testMixpanel)
-
-    let events = eventQueue(token: testMixpanel.apiToken)
-    XCTAssertEqual(events.count, 1)
-
-    let event = events.first
-    XCTAssertEqual(event?["event"] as? String, "$mp_page_view")
-
-    let props = event?["properties"] as? InternalProperties
-    XCTAssertEqual(props?["current_page_title"] as? String, "HomeScreen")
-
-    removeDBfile(testMixpanel.apiToken)
-  }
-
-  func testScreenViewNilScreenName() {
-    let testMixpanel = Mixpanel.initialize(token: randomId(), trackAutomaticEvents: false)
-
-    testMixpanel.screenView(screenName: nil)
-    testMixpanel.screenView(screenName: "")
-
-    waitForTrackingQueue(testMixpanel)
-
-    let events = eventQueue(token: testMixpanel.apiToken)
-    XCTAssertEqual(events.count, 0)
-
-    removeDBfile(testMixpanel.apiToken)
-  }
-
-  func testScreenLeave() {
-    let testMixpanel = Mixpanel.initialize(token: randomId(), trackAutomaticEvents: false)
-    testMixpanel.screenLeave(screenName: "HomeScreen", properties: ["time_spent": 30])
-
-    waitForTrackingQueue(testMixpanel)
-
-    let events = eventQueue(token: testMixpanel.apiToken)
-    XCTAssertEqual(events.count, 1)
-
-    let event = events.first
-    XCTAssertEqual(event?["event"] as? String, "$mp_page_leave")
-
-    let props = event?["properties"] as? InternalProperties
-    XCTAssertEqual(props?["current_page_title"] as? String, "HomeScreen")
-    XCTAssertEqual(props?["time_spent"] as? Int, 30)
-
-    removeDBfile(testMixpanel.apiToken)
-  }
-
-  func testScreenLeaveNilScreenName() {
-    let testMixpanel = Mixpanel.initialize(token: randomId(), trackAutomaticEvents: false)
-
-    testMixpanel.screenLeave(screenName: nil)
-    testMixpanel.screenLeave(screenName: "")
-
-    waitForTrackingQueue(testMixpanel)
-
-    let events = eventQueue(token: testMixpanel.apiToken)
-    XCTAssertEqual(events.count, 0)
-
-    removeDBfile(testMixpanel.apiToken)
   }
 
   func assertDefaultPeopleProperties(_ properties: InternalProperties) {
